refactor: extract signal event handlers from app.rs into handlers/signal.rs
Move handle_signal_event + 17 sibling handlers out of impl App into a
new src/handlers/signal.rs as free functions taking (&mut App, ...).
The shape follows the prior handlers/input.rs extraction from #398.

- src/app.rs loses 1276 lines, gains an 18-line shim plus a few
  visibility bumps (show_desktop_notification, path_to_file_uri,
  queue_single_read_receipt, PASTE_CLEANUP_DELAY_SECS)
- src/handlers/signal.rs is 1270 lines, all handler bodies
- 535 tests still passing, App field count baseline unchanged (66/66)
- handle_message's internal push_msg closure now takes app: &mut App
  as its first parameter; logic untouched

The 3-way handle_message split (resolve_incoming / push_resolved /
apply_notification_policy) suggested in #399 is left for a follow-up
PR since it is a behaviour-adjacent restructure on top of this purely
mechanical move.

Refs #399
